# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Campus GatePass System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, warden, security, admin]
        - name: search
          in: query
          schema:
            type: string
        - name: hostelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /hostels:
    get:
      operationId: listHostels
      tags: [hostels]
      summary: List all hostels
      responses:
        "200":
          description: List of hostels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hostel"
    post:
      operationId: createHostel
      tags: [hostels]
      summary: Create a hostel (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hostel"
  /hostels/{id}:
    patch:
      operationId: updateHostel
      tags: [hostels]
      summary: Update a hostel (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hostel"
    delete:
      operationId: deleteHostel
      tags: [hostels]
      summary: Delete a hostel (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, warden, security, admin]
        rollNumber:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        roomNumber:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
        hostelId:
          type: ["integer", "null"]
        hostelName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, warden, security, admin]
        rollNumber:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hostelId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        roomNumber:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        hostelId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        roomNumber:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    Hostel:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        totalRooms:
          type: ["integer", "null"]
        wardenId:
          type: ["integer", "null"]
        wardenName:
          type: ["string", "null"]
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    HostelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        totalRooms:
          type: ["integer", "null"]
        wardenId:
          type: ["integer", "null"]
    HostelUpdate:
      type: object
      properties:
        name:
          type: string
        totalRooms:
          type: ["integer", "null"]
        wardenId:
          type: ["integer", "null"]
...
    Notification:
      type: object
      required: [id, userId, title, message, type, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [gate_pass_approved, gate_pass_rejected, gate_pass_applied, late_return, general]
        isRead:
          type: boolean
        gatePassId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...