# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus GatePass System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  hostelId?: number | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  roomNumber?: string | null;
  /** @nullable */
  parentPhone?: string | null;
  isActive?: boolean;
}